Fixed map and unitRadius for plan request.

+++ server/src/main/java/com/tripco/t10/planner/Vincenty.java
@@ -1,10 +1,13 @@
 package com.tripco.t10.planner;
 
 /**
- * Calculates distance using vincenty formula.
+ * Called from child classes to calculate distance.
  */
 public class Vincenty{
-  public int calculateDistance(double[] coordinates, String units, int unitRadius){
+    /**
+     * Calculates distance using vincenty formula.
+     */
+  public int calculateDistance(double[] coordinates, String units, Double unitRadius){
   
     //Convert to Radians 
     double x1 = Math.toRadians(coordinates[0]);
